refactor: ActivityResult 관련 로직을 액티비티로 이동
