+++ app/src/main/java/woowacourse/payments/ui/cardlist/CardListScreen.kt
@@ -1,11 +1,5 @@
 package woowacourse.payments.ui.cardlist
 
-import android.app.Activity.RESULT_OK
-import android.content.Context
-import android.content.Intent
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.ActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -32,34 +26,21 @@ import woowacourse.payments.domain.CardNumber
 import woowacourse.payments.domain.CardholderName
 import woowacourse.payments.domain.ExpirationDate
 import woowacourse.payments.domain.Passcode
-import woowacourse.payments.ui.addcard.AddCardActivity
-import woowacourse.payments.ui.common.ExtraKeys
 import woowacourse.payments.ui.common.composable.PaymentCard
-import woowacourse.payments.ui.common.getParcelableExtraCompat
-import woowacourse.payments.ui.format.ExpirationDateFormat
 import woowacourse.payments.ui.model.CardUiModel
 import woowacourse.payments.ui.model.toUiModel
 import java.time.YearMonth
 
 @Composable
-fun CardListScreen(cards: SnapshotStateList<CardUiModel>) {
-    val context: Context = LocalContext.current
-    val launcher =
-        rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
-            if (result.resultCode == RESULT_OK) {
-                result.data?.toCardOrNull()?.let { card: Card ->
-                    cards.add(card.toUiModel())
-                }
-            }
-        }
-
-    fun navigateToAddCard() {
-        launcher.launch(AddCardActivity.intent(context))
-    }
+fun CardListScreen(
+    cards: SnapshotStateList<CardUiModel>,
+    onNavigateToAddCard: () -> Unit,
+) {
+    LocalContext.current
 
     Scaffold(
         modifier = Modifier.fillMaxSize(),
-        topBar = { CardListTopBar(cards) { navigateToAddCard() } },
+        topBar = { CardListTopBar(cards) { onNavigateToAddCard() } },
     ) { innerPadding: PaddingValues ->
         Column(
             horizontalAlignment = Alignment.CenterHorizontally,
@@ -86,32 +67,16 @@ fun CardListScreen(cards: SnapshotStateList<CardUiModel>) {
             }
 
             if (cards.size <= 1) {
-                AddCardButton { navigateToAddCard() }
+                AddCardButton { onNavigateToAddCard() }
             }
         }
     }
 }
 
-private fun Intent.toCardOrNull(): Card? =
-    runCatching {
-        getParcelableExtraCompat<CardUiModel>(ExtraKeys.CARD_KEY)?.let { card: CardUiModel ->
-            val yearMonth: YearMonth =
-                YearMonth.parse(card.expirationDate, ExpirationDateFormat.formatPattern)
-
-            Card(
-                CardNumber(card.cardNumber),
-                ExpirationDate(yearMonth),
-                CardholderName(card.cardholderName),
-                Passcode(card.passcode),
-                card.cardCompany.company,
-            )
-        }
-    }.getOrNull()
-
 @Preview(showBackground = true, name = "카드 목록 (0개)")
 @Composable
 private fun CardListScreenWithNoCardsPreview() {
-    CardListScreen(remember { mutableStateListOf() })
+    CardListScreen(remember { mutableStateListOf() }, {})
 }
 
 @Preview(showBackground = true, name = "카드 목록 (1개)")
@@ -129,6 +94,7 @@ private fun CardListScreenWithOneCardPreview() {
                 ).toUiModel(),
             )
         },
+        {},
     )
 }
 
@@ -136,24 +102,24 @@ private fun CardListScreenWithOneCardPreview() {
 @Composable
 private fun CardListScreenWithTwoCardsPreview() {
     CardListScreen(
-        cards =
-            remember {
-                mutableStateListOf(
-                    Card(
-                        CardNumber("1234123412341234"),
-                        ExpirationDate(YearMonth.of(2034, 12)),
-                        CardholderName("디랙"),
-                        Passcode("1234"),
-                        CardCompany.BC_CARD,
-                    ).toUiModel(),
-                    Card(
-                        CardNumber("1234123412341234"),
-                        ExpirationDate(YearMonth.of(2034, 12)),
-                        CardholderName("디랙"),
-                        Passcode("1234"),
-                        CardCompany.BC_CARD,
-                    ).toUiModel(),
-                )
-            },
+        remember {
+            mutableStateListOf(
+                Card(
+                    CardNumber("1234123412341234"),
+                    ExpirationDate(YearMonth.of(2034, 12)),
+                    CardholderName("디랙"),
+                    Passcode("1234"),
+                    CardCompany.BC_CARD,
+                ).toUiModel(),
+                Card(
+                    CardNumber("1234123412341234"),
+                    ExpirationDate(YearMonth.of(2034, 12)),
+                    CardholderName("디랙"),
+                    Passcode("1234"),
+                    CardCompany.BC_CARD,
+                ).toUiModel(),
+            )
+        },
+        {},
     )
 }
